Move button to dedicated directory
