+++ components/Destination.js
@@ -1,7 +1,7 @@
 import { Dialog } from "@headlessui/react"
 import { useState } from "react"
 
-import Button from "@components/Button"
+import Button from "@components/Button/Button/Button"
 import Card from "@components/Card"
 import Toggle from "@components/Toggle"
