+++ components/Icons/DataImport.js
@@ -1,0 +1,12 @@
+export function CentralDataImportIcon(props, fill) {
+  return (
+    <svg {...props} width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M6 0C6.4142 0 6.75 0.33579 6.75 0.75V5.25H11.25C11.6642 5.25 12 5.5858 12 6C12 6.4142 11.6642 6.75 11.25 6.75H6.75V11.25C6.75 11.6642 6.4142 12 6 12C5.5858 12 5.25 11.6642 5.25 11.25V6.75H0.75C0.33579 6.75 0 6.4142 0 6C0 5.5858 0.33579 5.25 0.75 5.25H5.25V0.75C5.25 0.33579 5.5858 0 6 0Z"
+        fill={fill}
+      />
+    </svg>
+  )
+}
