+++ components/SyncObject.js
@@ -1,6 +1,6 @@
 import { useState } from "react"
 
-import Button from "@components/Button"
+import Button from "@components/Button/Button/Button"
 import Card from "@components/Card"
 import RequestTooltip from "@components/RequestTooltip"
 import SyncEditWizard from "@components/SyncEditWizard"
