+++ components/Workflows/NewSourceFlow/NewSourceDrawer.js
@@ -1,6 +1,6 @@
 import Image from "next/image"
 
-import Button from "@components/Button"
+import Button from "@components/Button/Button/Button"
 import { useSourceFlow } from "@components/Contexts/SourceFlowContext"
 import {
   Drawer,
